Added a public "skipToMenuInit()" function to redraw the skipTo menu
Currently if headings, landmarks, etc are added via AJAX or DOM
scripting, these new items are not added to the skipTo menu.  As part
of this commit, now a skipToMenuInit() function can be called after DOM
changes to redraw the menu to incorporate the newly added items.

The ordering of the compiled javascript was necessary in order to
insure that the dropMenu function was called properly.